feat: repository 등록 API

See merge request s11-final/S11P31C206!39

+++ edith-back/developmentassistant/src/main/java/com/edith/developmentassistant/client/GitLabClient.java
@@ -16,7 +16,7 @@ public class GitLabClient {
 
     private final RestTemplate restTemplate;
 
-    public void registerWebhook(String branch, Integer projectId, String personalAccessToken) {
+    public void registerWebhook(String branch, Long projectId, String personalAccessToken) {
         String url = "https://lab.ssafy.com/api/v4/projects/" + projectId + "/hooks";
 
         // 헤더 설정
